refactor: move NewSiteCommand's logic onto Site, extract admin_auth
Site.provision() now owns new-site validation, provider registration,
app installation, pilot-comms config, asset building, hosts entry, and
SSL cert issuance; NewSiteCommand is a thin CLI wrapper around it.

Pulled forward the generate_session.py JWT-helper extraction (plan step
9) since Site.provision() needs ensure_jwt_secret/issue_site_token and
importing them from pilot.commands would have been a fresh layering
violation. Moved to pilot/core/admin_auth.py; generate_session.py and
every admin/backend caller now import from there.
