+++ admin/backend/api/v1/core.py
@@ -121,7 +121,7 @@ def create_session():
     if error is not None:
         return error
 
-    from pilot.commands.admin.generate_session import ensure_jwt_secret, issue_token
+    from pilot.core.admin_auth import ensure_jwt_secret, issue_token
 
     response = created_response(
         {"authenticated": True, "scope": "bench"},
