+++ tests/pilot/commands/test_commands.py
@@ -319,11 +319,11 @@ def test_new_command_skips_offset_with_live_admin_internal_port(tmp_path: Path, 
     assert data["bench"]["http_port"] == 8002
 
 
-# ── NewSiteCommand ────────────────────────────────────────────────────────────
+# ── Site.provision validation ────────────────────────────────────────────────
 
 
 def test_new_site_raises_if_site_exists(tmp_path: Path) -> None:
-    from pilot.commands.sites.create import NewSiteCommand
+    from pilot.core.site import _validate_new_site
 
     bench = make_bench(tmp_path)
     bench.create_directories()
@@ -332,41 +332,42 @@ def test_new_site_raises_if_site_exists(tmp_path: Path) -> None:
     (site_dir / "site_config.json").write_text("{}")
 
     with pytest.raises(BenchError, match="already exists"):
-        NewSiteCommand(bench, "site1.localhost", ["frappe"], "secret")._validate()
+        _validate_new_site(bench, "site1.localhost", ["frappe"])
 
 
 def test_new_site_raises_if_app_not_in_apps_txt(tmp_path: Path) -> None:
-    from pilot.commands.sites.create import NewSiteCommand
+    from pilot.core.site import _validate_new_site
 
     bench = make_bench(tmp_path)
     bench.create_directories()
     (bench.sites_path / "apps.txt").write_text("frappe\n")
 
     with pytest.raises(BenchError, match="erpnext"):
-        NewSiteCommand(bench, "site1.localhost", ["erpnext"], "secret")._validate()
+        _validate_new_site(bench, "site1.localhost", ["erpnext"])
 
 
 def test_new_site_validate_passes_when_all_ok(tmp_path: Path) -> None:
-    from pilot.commands.sites.create import NewSiteCommand
+    from pilot.core.site import _validate_new_site
 
     bench = make_bench(tmp_path)
     bench.create_directories()
     (bench.sites_path / "apps.txt").write_text("frappe\n")
 
-    NewSiteCommand(bench, "site1.localhost", ["frappe"], "secret")._validate()  # no raise
+    _validate_new_site(bench, "site1.localhost", ["frappe"])  # no raise
 
 
 def test_new_site_validate_passes_with_no_apps_requested(tmp_path: Path) -> None:
-    from pilot.commands.sites.create import NewSiteCommand
+    from pilot.core.site import _validate_new_site
 
     bench = make_bench(tmp_path)
     bench.create_directories()
 
-    NewSiteCommand(bench, "site1.localhost", [], "secret")._validate()  # no raise
+    _validate_new_site(bench, "site1.localhost", [])  # no raise
 
 
 def test_build_missing_assets_skips_cloned_but_unregistered_apps(tmp_path: Path) -> None:
-    from pilot.commands.sites.create import NewSiteCommand
+    from pilot.config.site import SiteConfig
+    from pilot.core.site import Site
 
     bench = make_bench(tmp_path)
     bench.create_directories()
@@ -378,7 +379,7 @@ def test_build_missing_assets_skips_cloned_but_unregistered_apps(tmp_path: Path)
     with patch(
         "pilot.managers.python_environment.PythonEnvManager.build_assets_for_app"
     ) as build:
-        NewSiteCommand(bench, "site1.localhost", ["frappe"], "secret").build_missing_assets()
+        Site(SiteConfig(name="site1.localhost", apps=["frappe"]), bench)._build_missing_assets()
 
     built = {call.args[0].config.name for call in build.call_args_list}
     assert built == {"frappe"}
